refactor: add sse helpers and use in local search
